+++ @types/index.d.ts
@@ -1,8 +1,3 @@
-declare module '*.mdx' {
-  const content: React.FunctionComponent<{}>;
-  export default content;
-}
-
 declare module '*.module.css' {
   const classes: { [key: string]: string };
   export default classes;
@@ -10,3 +5,7 @@ declare module '*.module.css' {
 
 declare module '@fabric-ds/css';
 declare module '@fabric-ds/css/tailwind-css';
+
+declare module '@warp-ds/icons/react' {
+  export * from '@warp-ds/icons/dist/types/react'
+}