+++ packages/_helpers/affix.tsx
@@ -1,15 +1,11 @@
-import React from 'react';
-import { suffix, prefix } from '@warp-ds/css/component-classes';
-import { classNames } from '@chbphone55/classnames';
-import { i18n } from '@lingui/core';
-import { messages as enMessages} from './locales/en/messages.mjs';
-import { messages as nbMessages} from './locales/nb/messages.mjs';
-import { messages as fiMessages} from './locales/fi/messages.mjs';
-import { activateI18n } from '../i18n';
+import React from "react";
+import { suffix, prefix } from "@warp-ds/css/component-classes";
+import { classNames } from "@chbphone55/classnames";
+import { IconClose16, IconSearch16 } from "@warp-ds/icons/react";
 
 interface AffixProps {
   /** Defines a string value that labels the affix element. */
-  'aria-label'?: string;
+  "aria-label"?: string;
 
   /** Affix added at the beginning of input */
   prefix?: boolean;
@@ -32,19 +28,12 @@ interface AffixProps {
 
 export function Affix(props: AffixProps) {
   const classBase = props.prefix ? prefix : suffix;
-  activateI18n(enMessages, nbMessages, fiMessages);
-  const ariaLabelMagnifyingGlass = i18n._(
-    /*i18n*/ {
-      id: 'forms.affix.aria.magnifyingGlass',
-      message: 'Magnifying glass',
-      comment: 'Aria label for magnifying glass used as affix in Combobox and Textfield components',
-    });
 
   return React.createElement(
-    props.label ? 'div' : 'button',
+    props.label ? "div" : "button",
     {
-      'aria-label': !props.label ? props['aria-label'] : undefined,
-      type: props.search ? 'submit' : props.clear ? 'reset' : undefined,
+      "aria-label": !props.label ? props["aria-label"] : undefined,
+      type: props.search ? "submit" : props.clear ? "reset" : undefined,
       onClick: props.onClick,
       className: classNames({
         [classBase.wrapper]: true,
@@ -53,53 +42,11 @@ export function Affix(props: AffixProps) {
       }),
     },
     <>
-      {props.clear && (
-        <svg
-          role="img"
-          aria-label="X"
-          xmlns="http://www.w3.org/2000/svg"
-          width="16"
-          height="16"
-          fill="none"
-          viewBox="0 0 16 16"
-        >
-          <path
-            fill="currentColor"
-            fillRule="evenodd"
-            d="M4.03 2.97a.75.75 0 00-1.06 1.06L6.94 8l-3.97 3.97a.75.75 0 101.06 1.06L8 9.06l3.97 3.97a.75.75 0 101.06-1.06L9.06 8l3.97-3.97a.75.75 0 00-1.06-1.06L8 6.94 4.03 2.97z"
-            clipRule="evenodd"
-          />
-        </svg>
-      )}
+      {props.clear && <IconClose16 />}
 
-      {props.search && (
-        <svg
-          role="img"
-          aria-label={ariaLabelMagnifyingGlass}
-          xmlns="http://www.w3.org/2000/svg"
-          width="16"
-          height="16"
-          fill="none"
-          viewBox="0 0 16 16"
-        >
-          <g
-            stroke="currentColor"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth="1.5"
-            clipPath="url(#nra-cclip0)"
-          >
-            <path d="M8.796 11.803A5.684 5.684 0 104.349 1.341a5.684 5.684 0 004.447 10.462zM11 11l4 4" />
-          </g>
-          <defs>
-            <clipPath id="nra-cclip0">
-              <path fill="currentColor" d="M0 0h16v16H0z" />
-            </clipPath>
-          </defs>
-        </svg>
-      )}
+      {props.search && <IconSearch16 />}
 
       {props.label && <span className={classBase.label}>{props.label}</span>}
-    </>,
+    </>
   );
 }