+++ packages/pill/src/component.tsx
@@ -59,19 +59,7 @@ export function Pill(props: PillProps) {
                 },
               )}
           </span>
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            width="12"
-            height="12"
-            fill="none"
-            viewBox="0 0 16 16"
-            aria-hidden="true"
-          >
-            <path
-              fill="currentColor"
-              d="M4.03 2.97a.75.75 0 00-1.06 1.06L6.94 8l-3.97 3.97a.75.75 0 101.06 1.06L8 9.06l3.97 3.97a.75.75 0 101.06-1.06L9.06 8l3.97-3.97a.75.75 0 00-1.06-1.06L8 6.94 4.03 2.97z"
-            ></path>
-          </svg>
+          <span>x</span>
         </button>
       )}
     </div>