+++ vite.config.js
@@ -1,14 +1,7 @@
-import mdx from 'vite-plugin-mdx';
 import react from '@vitejs/plugin-react';
-import slug from 'remark-slug';
-import autoLinkHeadings from 'remark-autolink-headings';
-import jsxExample from './docs/src/plugins/jsxExample.cjs';
-import propTable from './docs/src/plugins/propTable.cjs';
-import { presetWarp } from '@warp-ds/uno'
-import uno from 'unocss/vite'
+import { presetWarp } from '@warp-ds/uno';
+import uno from 'unocss/vite';
 import { classes } from '@warp-ds/css/component-classes/classes';
-
-
 
 export default function config() {
   return {
@@ -19,12 +12,6 @@ export default function config() {
         safelist: classes,
       }),
       react(),
-      mdx.default({
-        remarkPlugins: [slug, autoLinkHeadings, jsxExample, propTable],
-      }),
     ],
-    build: {
-      outDir: 'site',
-    },
   };
 }